% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_PMScanR.R
\name{build_server}
\alias{build_server}
\title{Create the server function for PMScanR}
\usage{
build_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
A Shiny server function
}
\description{
Create the server function for PMScanR
}
