% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_PMScanR.R
\docType{data}
\name{build_ui}
\alias{build_ui}
\title{Create the main UI for PMScanR}
\format{
An object of class \code{bslib_page} (inherits from \code{shiny.tag.list}, \code{list}) of length 5.
}
\usage{
build_ui
}
\value{
A Shiny UI object
}
\description{
Create the main UI for PMScanR
}
\keyword{datasets}
