% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_segments.R
\name{extract_segments}
\alias{extract_segments}
\title{Extract sequences from a file within a given range}
\usage{
extract_segments(seq, from, to)
}
\arguments{
\item{seq}{A FASTA file where are a sequences}

\item{from}{An integer specifying the starting position to extract from the file.}

\item{to}{An integer specifying the ending position to extract from the file}
}
\value{
A list which are representing the extracted sequences.
         Sequences shorter than the specified range are excluded.
}
\description{
This function extracts sequences from a file within the specified range (`from` to `to`).
Sequences shorter than the specified range are excluded from the result.
}
\examples{
seq <- system.file(file = "extdata/hemoglobins.fasta", package = "PMScanR")
from = 10
to = 20
if (seq != "") {
segment_motifs <- extract_segments(seq, from = from, to = to)
}
}
