% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff2matrix.R
\name{gff2matrix}
\alias{gff2matrix}
\title{Convert GFF to a matrix}
\usage{
gff2matrix(input)
}
\arguments{
\item{input}{A data frame containing GFF data.
Typically, this data frame is generated by importing a GFF file using `rtracklayer::import.gff`
and converting it to a data frame}
}
\value{
A matrix, where values are binary: `1` indicates the presence of a feature in a sequence,
         and `0` indicates its absence
}
\description{
This function takes a GFF data frame and converts it into a matrix
}
\examples{
Load required libraries
library(rtracklayer)
library(reshape2)

gff_file <- system.file("extdata", "example.gff", package = "rtracklayer")
gff_data <- rtracklayer::import.gff(gff_file)
gff_data <- as.data.frame(gff_data)

matrix <- gfftomatrix(gff_data)
print(matrix)

}
