% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMScanR-package.R
\docType{package}
\name{PMScanR-package}
\alias{PMScanR}
\alias{PMScanR-package}
\title{PMScanR: Protein motifs analysis and visualisation}
\description{
Provides tools for large-scale protein motif analysis and visualization in R. PMScanR facilitates the identification of motifs using external tools like PROSITE's ps_scan (handling necessary file downloads and execution) and enables downstream analysis of results. Key features include parsing scan outputs, converting formats (e.g., to GFF-like structures), generating motif occurrence matrices, and creating informative visualizations such as heatmaps, sequence logos (via seqLogo/ggseqlogo). The package also offers an optional Shiny-based graphical user interface for interactive analysis, aiming to streamline the process of exploring motif patterns across multiple protein sequences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/prodakt/PMScanR}
  \item Report bugs at \url{https://github.com/prodakt/PMScanR/issues}
}

}
\author{
\strong{Maintainer}: Jan Pawel Jastrzebski \email{bioinformatyka@gmail.com} (\href{https://orcid.org/0000-0001-8699-7742}{ORCID})

Authors:
\itemize{
  \item Damian Czopek \email{dcwmpl@gmail.com} (\href{https://orcid.org/0009-0005-3471-4866}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Monika Gawronska \email{gawronska572@gmail.com} (\href{https://orcid.org/0009-0001-2677-6371}{ORCID}) [contributor]
  \item Wiktor Babis \email{wiktorbabis@gmail.com} (\href{https://orcid.org/0009-0006-3648-3413}{ORCID}) [contributor]
  \item Miriana Quaranta \email{miriana.quaranta@uniroma1.it} (\href{https://orcid.org/0009-0003-0855-485X}{ORCID}) [contributor]
}

}
\keyword{internal}
