% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_segments.R
\name{prepare_segments}
\alias{prepare_segments}
\title{Extract a fragment from a file within a given range}
\usage{
prepare_segments(seq, from, to)
}
\arguments{
\item{seq}{A character vector representing a sequence from file}

\item{from}{An integer specifying the starting position of the fragment to extract}

\item{to}{An integer specifying the ending position of the fragment to extract}
}
\value{
A character string representing the extracted subsequence.
         If the sequence is shorter than the specified range, `NA` is returned
}
\description{
This function extracts a sequence from a given file based on the specified range (`from` to `to`).
If the sequence is shorter than the specified range, the function returns `NA`
}
\examples{
seq <- read.fasta(file = 'file', seqtype = "AA")
fragment <- prepare_segments(seq, from = 10, to = 20)
print(fragment)

}
