% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.prosite.R
\name{read.prosite}
\alias{read.prosite}
\title{Convert PROSITE format to GFF}
\usage{
read.prosite(prosite_input)
}
\arguments{
\item{prosite_input}{A file in PROSITE format from Prosite database}
}
\value{
A GFF file as data frame.
}
\description{
This function converts data in PROSITE format into a GFF data frame.
PROSITE is a database of protein families and domains, and this function parses the relevant information
into a structured GFF data frame suitable for further analysis and visualization.
}
\examples{
prosite_file <- system.file("inst/extdata/PROSITEoutput.txt", package = "PMScanR")
if (prosite_file != "") {
gff_data <- read.prosite(prosite_file)
}

}
