% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqPie.R
\name{freqPie}
\alias{freqPie}
\title{Create a pie chart from a GFF file which showing the distribution of protein motifs in the 'Name' column}
\usage{
freqPie(data)
}
\arguments{
\item{data}{A file in GFF format containing a column named 'Name' with the names of each protein motifs}
}
\value{
A pie chart in ggplot which displays the percentage distribution of each protein motifs
        in the `Name` column, with labels showing the protein motifs name and its corresponding percentage.
 @example
data <- 'GFF file'
count <- name_counts(data)
piechart <- ggplot(name_counts, aes(x = "", y = count, fill = Name_and_procent)) +
             geom_bar(stat = "identity", width = 1) +
             coord_polar(theta = "y") +
             theme_minimal() +
             scale_fill_brewer(palette = "Set1") +    # Optional colour selection to be included
             theme(axis.text.x = element_blank())
}
\description{
This function calculates the number of occurrences and percentages for each protein motifs in the `Name` column.
It then creates a pie chart using `ggplot2` to visualize the distribution of these protein motifs.
}
