% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.psa.R
\name{read.psa}
\alias{read.psa}
\title{Convert PSA to GFF}
\usage{
read.psa(psa_input)
}
\arguments{
\item{psa}{A character vector where each element represents a line from a PSA file, file in a psa format}
}
\value{
A data frame in GFF format
}
\description{
This function converts a PSA format data into a GFF format data frame
}
\examples{
raw_data <- 'file'
s1 <- PSAtoGFF(raw_data)
print(raw_data)

}
