% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.psa.R
\name{read.psa}
\alias{read.psa}
\title{Convert PSA to GFF}
\usage{
read.psa(psa_input)
}
\arguments{
\item{psa_input}{A character vector where each element represents a line from a PSA file, file in a psa format}
}
\value{
A data frame in GFF format
}
\description{
This function converts a PSA format data into a GFF format data frame
}
\examples{
file_path <- system.file("extdata/out_Hb_psa.txt", package = "PMScanR")
if (file_path != "") {
  gffdata <- read.psa(file_path)
}
}
