% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runPMScanRShiny}
\alias{runPMScanRShiny}
\title{Launch the PMScanR Shiny app}
\usage{
runPMScanRShiny()
}
\value{
Launches the GUI for users not familiar with R and those who prefer graphical interface.
}
\description{
Launch the PMScanR Shiny app
}
\examples{
runPMScanRShiny()
}
