% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPsScan.R
\name{runPsScan}
\alias{runPsScan}
\title{Run PS-Scan}
\usage{
runPsScan(
  in_file,
  out_file,
  out_format,
  ps_scan = NULL,
  patterns_dat = NULL,
  pf_scan = NULL,
  OS = NULL
)
}
\arguments{
\item{in_file}{Input file containing protein sequences.}

\item{out_file}{A character string specifying the path to the output file where results will be saved.}

\item{out_format}{Specifying the output format (e.g., "fasta").}

\item{ps_scan}{A character string specifying the path to the PS-Scan Perl script. If NULL, it will be downloaded.}

\item{patterns_dat}{A character string specifying the path to the PROSITE patterns database file. If NULL, it will be downloaded.
This file can be downloaded from: https://ftp.expasy.org/databases/prosite/}

\item{pf_scan}{A character string specifying the path to the pfscan executable. If NULL, it will be downloaded and extracted based on OS (except for MAC).}

\item{OS}{Operating system ("WIN", "LINUX", "MAC"). If NULL, it will be detected automatically.}
}
\value{
Writes the results of the PS-Scan analysis to the specified output file.
}
\description{
Extended function to run PS-Scan with automatic OS detection, file downloading, extraction, and improved error handling.
}
\examples{
{
if (interactive()) {
  in_file <- system.file("extdata", "hemoglobins.fasta", package = "PMScanR")
  out_file <- tempfile(fileext = ".gff")
  runPsScan(in_file = in_file, out_format = 'gff', out_file = out_file)
}
}
}
